move connection from menu onto update tail page
